Refactor file parser to separated component
